Fix: follow 유저타입 수정

+++ backend/raonzena/src/main/java/com/ssafy/raonzena/api/response/UserProfileRes.java
@@ -18,7 +18,7 @@ import java.sql.Timestamp;
 @AllArgsConstructor
 public class UserProfileRes {
 
-    private Integer userNo;
+    private long userNo;
 
     private String userId;
